requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
PyMySQL==1.1.1